@@ -3,7 +3,7 @@ package com.poc.user.domain.request;
 
 import lombok.Data;
 
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotBlank;
 
 /**
  * Class that represents a user request body for particular user. Same fields as  {@link UserInfo UserInfo.class}
